# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CasperProver API
  description: |
    ZK-proof engine for verifiable AI decisions on Casper Network.
    32 endpoints covering proof generation, ZK verification, post-quantum
    cryptography, batch aggregation, KYC gating, and proof-chain DAG validation.
  version: 0.4.0
  license:
    name: MPL-2.0
servers:
  - url: https://casperprover-api-ylsh.onrender.com
    description: Production (Casper testnet)
  - url: http://localhost:9090
    description: Local development
...
paths:
...
  /proofs:
    get:
      tags: [Proofs]
      summary: List proofs
      operationId: listProofs
      parameters:
        - name: agent
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Paginated proof list
    post:
      tags: [Proofs]
      summary: Create a proof
      operationId: submitProof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent, model_hash, input_data]
              properties:
                agent: { type: string }
                model_hash: { type: string }
                input_data: { type: string }
                use_case: { type: string }
      responses:
        '201':
          description: Created proof with hash, Merkle root, and deploy hash
...
  /verify:
    post:
      tags: [Proofs]
      summary: Verify a proof
      operationId: verifyProof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proof_id]
              properties:
                proof_id: { type: string }
      responses:
        '200':
          description: Verification result with Merkle path check
...